[♻️refactor]: Callout 컴포넌트 types와 theme 파일 분리
